converted SignupPage, LoginPage, NotFound to typescript. Created a index.d.ts file to support images in typescript
