+++ client/Components/App/App.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import NotFound from '../../Pages/NotFound/NotFound.jsx';
+import NotFound from '../../Pages/NotFound/NotFound';
 import HomePage from '../../Pages/Home/HomePage';
-import LoginPage from '../../Pages/Login/LoginPage.jsx';
-import SignupPage from '../../Pages/Signup/SignupPage.jsx';
+import LoginPage from '../../Pages/Login/LoginPage';
+import SignupPage from '../../Pages/Signup/SignupPage';
 import NavBar from '../NavBar/NavBar';
 import { io, Socket } from 'socket.io-client';
