+++ index.d.ts
@@ -1,0 +1,18 @@
+declare module '*.jpg' {
+  const jpgPath: string;
+  export default jpgPath;
+}
+declare module '*.jpeg' {
+  const jpegPath: string;
+  export default jpegPath;
+}
+declare module '*.svg' {
+  const svgPath: string;
+  export default svgPath;
+}
+declare module '*.png' {
+  const pngPath: string;
+  export default pngPath;
+}
+// to notify typescript of type for images. 
+
